Cleanup duplicate code for version specific function handleMessageChatOpcode
Fix channel message :-)
